Add calculation precision;

+++ app/src/main/java/com/example/calculaltorproject/domain/model/Actions.kt
@@ -2,6 +2,6 @@ package com.example.calculaltorproject.domain.model
 
 sealed interface Actions {
     data class Sign(val sign: String) : Actions
-    data class Number(val numberSign: String, val number:Long) : Actions
+    data class Number(val numberSign: String, val number: Double) : Actions
 
 }
